+++ sprout_app/lib/features/settings/presentation/settings_page.dart
@@ -3,9 +3,9 @@ import 'package:flutter_bloc/flutter_bloc.dart';
 import 'package:go_router/go_router.dart';
 import 'package:package_info_plus/package_info_plus.dart';
 import 'package:purchases_ui_flutter/purchases_ui_flutter.dart';
-import 'package:debug_lens/debug_lens.dart';
 
 import 'package:sprout/core/core.dart';
+import 'package:sprout/core/debug/sprout_debug_lens.dart';
 import 'package:sprout/features/auth/export.dart';
 import 'package:sprout/features/purchases/presentation/premium_paywall_helper.dart';
 import 'package:sprout/ui/export.dart';
@@ -149,7 +149,7 @@ class _SettingsPageState extends State<SettingsPage> {
   }
 
   void _openDebugLens() {
-    DebugLens.show(context);
+    SproutDebugLens.show(context);
   }
 
   @override